Add Clipboard Monitor.
Add code to handle threads in ClipboardInterfaceController.
Add an SimpleStringProperty to wrap around the String used
to fill the clipboard TextArea.
Add a thread to monitor the system clipboard, and update the
clipboard TextArea as needed.

Issues:  We are running a Task, which runs in a loop.  This loop
sleeps 100 ms, then passes a Runnable to Platform.runLater for
UI updates.  We have no control over when Platform.runLater will
actually run the Runnable, so updates may not always occur every
100 ms, and it is possible for Runnables to clog up the JavaFX
UI thread.  This "works" for now, but may not be the best approach.