Modify UI.  Add buffer TextArea.
Add a second TextArea.
Add a clear clipboard button.
Top TextArea is associated with the system clipboard and
displays the contents of the system clipboard.  It is updated
with changes via the clipboard monitor.
We have a button that clears the system clipboard.
The bottom text area is a general buffer.
Modified the read/write to clipboard handlers to work with the buffer
TextArea.  The idea is that the user can work on data in the buffer,
or store it to the side while copying some other value to the
clipboard.

The goal is to have more buffers giving the user more buffers to
store data that could be moved to the system clipboard.