Bugfix for nulls in the ClipboardMonitorRunnable.
Fixes a bug when running multiple copies of MultiClipboard.
When multiple copies of MultiClipboard are run,
the clipboard will return null when getString is called.

Get reference to the JavaFX System Clipboard in
ClipboardMonitorTask, and pass it in to ClipboardMonitorRunnable
via constructor.
This avoids us calling getSystemClipboard every 100 milliseconds.
Seems to fix this bug, but logging is in place in case this bug
comes up again in the future.