Quick workaround for data == null issue.
Sometimes Clipboard.hasString() returns true, and
Clipboard.getString() will return null.

Not sure what causes this issue, but adding a check in the if
condition at the top of run() keeps this bug from causing
problems for us.

Not a real fix, need to find out what conditions will cause
JavaFX Clipboard to behave in this odd manner.  This will lead us
to a real fix.