Fix a state management bug.
While nuking the clipboard, oldData did not get cleared out.

If it held the previous contents of the clipboard, and if the
buffer contains the previous contents of the clipboard, the user
will not be able to write the buffer contents to the clipboard since
the check on !this.oldData.equals(data) will fail.

Really, we don't need oldData, it is a premature optimization.