Fix clipboard contents binding.
Change behavior of DataTableRowSelectionListener, if newValue
BufferBase object is the clipboard, bind the bufferTextArea
to the clipboardContents SimpleStringProperty, disable editing.

When the current buffer is the clipboard we want bufferTextArea to be
bound to the clipboardContents SimpleStringProperty.

When the current buffer is not the clipboard, we don't want the
bufferTextArea to be bound to anything.  We want to set it the value of
the current buffer, and make bufferTextArea editable.

Add logic to change behavior depending on BufferBase newValue type.