Fix a regression
Accidentally removed a line that we needed to enable editing for
the name column in the Buffer Table View.

We had some complicated code in place to work around having the
clipboard as a buffer.  That code got removed when it was
decided that the clipboard buffer was no longer needed.

I must have nuked this line by accident when removing the clipboard
buffer related code.

Restored this line to fix name column editablity.